Fix Yaml date parsing issue
Hacky fix for yaml date parsing.  Symfony's yaml parser converts
dates to seconds-since-epoch...but ES uses milliseconds.  Since
the date is converted to a number, there isn't a good way to
determine if the field was originally a date or number.

This patch pre-parses the yaml and replaces dates with the appropriate
ms-since-epoch.

Will remove once upstream has been modified