Fix behavior when a string body is provided
Strings (and arrays) should simply be given to the serializer
as-is, without trying to guess the nature of the string.  Previously,
the endpoint would pack strings into the 'q' URI parameter, which
is opinonated and buggy.

This patch now just sets the body and lets the serializer figure
it out.

Closes #5