Add a Static Connection pool that does not ping
Implements StaticNoPingConnectionPool.  This class is used when
you have a static list of nodes and do not want any pinging or
health check.

Nodes are optimistically assumed to be alive.  If they fail a
request, they are marked dead until their timeout is expired.  If
there are no nodes left alive then the connection pool terminates
with an exception (unlike StaticConnectionPool which attempts one
final round of pings).

This class should be used for fail-fast situations that require
low overhead/latency.

Closes #21