Fix bug where retries were not being properly set
Stung by the infamous isset() vs null issue in PHP.  Need to check if
the `retries` param is set OR if it is null, since the default set
of config options set `retries` to null.

This bug prevented retries from taking place, which meant that under
normal operating scenarios a bad host could take down the client with
an exception, when it normally should have retried on a different host.