markAlive() and markDead() should also set lastPing time
The act of marking a node dead or alive means it was tried via a
normal API call.  Since some connection pools don't ping, we should
use markAlive/Dead as proxies for pinging to optimize cluster
load balancing (e.g. not repeatedly trying to call a dead node
because it doesn't have a ping timeout)