Update mechanism to retrieve yaml tests
Migrating away from a pure-composer solution because it is too
difficult to sync multiple branches/tests.  Added a new RestSpecRunner
which will checkout the appropriate commit/branch based on an env
variable.

In the future will add logic to also scan for locally running
clusters and find the appropriate tests.