[TEST] Use forked version of Symfony/Yaml
The forked Yaml parser supports map via the stdClass() object, which
is a required feature for a number of yaml tests.  This commit
switches over to the forked parser until the point in time which
Symfony devs decide to merge the PR.

Commit also includes various changes to the YamlTestRunner to make it
compatible with objects.