Fix bug where scheme prefix always defaulted to HTTP
The host parser correctly extracted scheme prefixes if they existed,
but the AbstractConnection always defaulted to plain HTTP.  This commit
adds a fix which allows the connection to set the scheme.

The default 'http' is retained for backwards compatibility. Should
be depreciated at some point.

Also adds a simple test and documentation

Closes to #75