Update indexing-operations.asciidoc
The example uses the 'doc' parameter on bulk indexing, which is inconsistent with how the bulk API expects JSON documents to be passed in, and is thus incorrect. It essentially puts all the data under a 'doc' property. Only the update action supports the 'doc' parameter.

I struggled with this for quite some time before realizing the problem :)