Rework JSON Serializer to Throw Useful Exceptions
Closes #175

Previously `SmartSerializer` was the only class that sometimes threw
exceptions when something went wrong during JSON deserialization. This
changes that so all serializers throw.

- `JsonSerializationError` is thrown when something goes wrong during
  `json_encode`
- `JsonDeserializationError` is thrown when something goes wrong during
  `json_decode`

Both are subclasses of `JsonErrorException`, so anyone already catching
that should be able to continue doing so.

There's a also a bit of restructuring here and some additional tests to
cover some more paths in the serializers.