[TEST] Invoke phpunit dependency instead of travis phpunit.phar
phpunit.phar bundles dependencies, and travis does not allow control over which version of
phpunit is used.  This means phpunit 5.x is used in the PHP 7.x branch, which unfortunately
bundles symfony/yaml 3.x.

Symfony/yaml 3.x has a breaking change that is incompatible with 2.x, so the tests are split
between 7.x failing and 5.x passing, or vice versa.

So we need to use the phpunit dependency instead of Travis.  At least until 5.x is phased out
and we can cutover to symfony/yaml 3.x across the branches.