[Internal BWC Break] Refactor to remove Transport dependence in endpoints
This removes the reliance on Transport inside of endpoints.  Instead of endpoints calling
transport->performRequest(), the endpoint is used as a temporary object that performs
validation and input manipulation.  The Namespace then uses this object to invoke
transport->performRequest() directly.

This will make injecting custom namespaces significantly easier, since it makes the dependency
graph easier to wire.  It also makes mocking / unit tests easier.

Related to #421