[Internal BWC Break] Add better ability to inject namespaces
Adds a new `registerNamespace` method to the client builder, and supporting infrastructure
to use newly registered namespaces in the client.  It was previously possible to do this
via the `setEndpoint()` method... but it was very difficult and not pleasant to use.

Now, users can implement the `NamespaceBuilderInterface` interface, which is registered into
the client build and used to instantiate user-defined namespaces when the client is built.
Those namespaces are then automatically used by the client as if they were built-in.

Note: namespaces cannot conflict with existing namespaces (if they do, they will never be called
since this uses the __call dynamic method, which will defer to explicit method names first).
Also note that two registered namespaces with the same name will results in a
last-register-wins situation (e.g. they are stored in a map), so avoid conflicts :)