If request data is not properly encoded the serialisers can return boolean false
This eventually leads to an uncaugh InvalidArgumentException:
"Invalid request body provided" from GuzzleHttp\Ring\Client\CurlFactory
which is expecting a body.

IMHO throwing an exception early would be preferable, as it should explain
why the body is not correct and make things easier for developers to find
their code which is not encoding the body correctly.