Remove incorrect Composer instructions from README
Instructions on using the --no-dev flag were wrong as Composer will in this case never install the dev dependencies anyway. Dev dependencies are only ever resolved in the root package of an application for exactly the reasons that were described here. A consumer of the library will therefore never get the dev requirements of any library, with any flag.