Append newline to bulk string requests, reject invalid types
If the user provides a string as the complete bulk request, it must end
with a newline for Elasticsearch to be happy.  This is something we can
check and fix for the user.

We should also throw an exception if the user provides anything other
than an array, traversable object or string.