Trim trailing slashes from hosts when passed as a string.
When a host contains trailing slashes this might lead to issues when requesting the path `/`,
as this might lead to the uri `//`. This is an invalid uri - `parse_url()` will return `false`
on this uri.

This PR only removes trailing slashes from hosts that are passed as a string. Rationale behind
this decision is that, when passed as an array, the path is a separate entry in that array.

Maybe this should be handled by the code that calls the `ClientBuilder` or maybe the handler
where this issue presented itself in this scenario, but I think by sanitizing the uri by
proactively removing the trailing slashes this library will become more robust.

This PR fixes #869